requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
tensorboard
future
tqdm